UML Activity Diagram Example 2
This is an example of a customer sending purchase requests to a Purchase Bid..
Please note you will be able to open this EA13 diagram only if you have a licensed version or a trial version!



 Caseworkers process requests for the purchase of products
under $1,500, write a purchase order, and then send it to
the approved vendor.  Purchase requests over $1,500 must
first be sent out for bid from vendors that supply the
product.  When the bids return, the case worker selects one
bid.  Then, he or she writes a purchase order and sends it
to a vendor selected.